RP2040 MBED compile warnings

+++ src/AudioTimer/AudioTimerDef.h
@@ -40,7 +40,7 @@ class AudioUtils {
         static uint32_t toTimeUs(uint32_t samplingRate, uint8_t limit=10){
             uint32_t result = 1000000l / samplingRate;
             if (result <= limit){
-                LOGW("Time for samplingRate %u -> %u is < %u μs - we rounded up", samplingRate, result, limit);
+                LOGW("Time for samplingRate %u -> %lu is < %u μs - we rounded up", (unsigned int)samplingRate, result, limit);
                 result = limit;
             }
             return result;
@@ -49,7 +49,7 @@ class AudioUtils {
         static uint32_t toTimeMs(uint32_t samplingRate, uint8_t limit=1){
             uint32_t result = 1000l / samplingRate;
             if (result <= limit){
-                LOGW("Time for samplingRate %u -> %u is < %u μs - we rounded up", samplingRate, result, limit);
+                LOGW("Time for samplingRate %u -> %lu is < %u μs - we rounded up", (unsigned int)samplingRate, result, limit);
                 result = limit;
             }
             return result;
